Start migration of own namespace, coding style ...
