+++ include/jbr/reg/config.hpp.in
@@ -1,15 +1,15 @@
 //!
-//! @file include/jbr/reg/config.hpp.in
+//! @file jbr/reg/config.hpp.in
 //! @author jbruel
 //! @date 29/04/19
 //!
 
-#ifndef JBR_CREGISTER_CONFIG_HPP
-# define JBR_CREGISTER_CONFIG_HPP
+#ifndef CREGISTER_JBR_REG_CONFIG_HPP
+# define CREGISTER_JBR_REG_CONFIG_HPP
 
 # define VERSION_MAJOR "@cRegister_VERSION_MAJOR@"
 # define VERSION_MINOR "@cRegister_VERSION_MINOR@"
 # define VERSION_PATCH "@cRegister_VERSION_PATCH@"
 # define VERSION "@cRegister_VERSION@"
 
-#endif //JBR_CREGISTER_CONFIG_HPP
+#endif // !CREGISTER_JBR_REG_CONFIG_HPP
