+++ src/reg/library.cpp
